+++ src/base/Link.php
@@ -128,6 +128,7 @@ abstract class Link extends Element implements LinkInterface
     public bool $isFieldRequired = false;
 
     private ?FieldLayout $_fieldLayout = null;
+    private ?string $_text = null;
 
 
     // Public Methods
@@ -394,7 +395,18 @@ abstract class Link extends Element implements LinkInterface
 
     public function getText(?string $text = null): ?string
     {
+        // Allow custom text (defined at render) to override everything. This allows us to reference `linkText`
+        // which is the "original" link text, and `text` which is the derivative text.
+        if ($this->_text) {
+            return $this->_text; 
+        }
+
         return $this->getLinkText() ?: trim($this->getLinkUrl() . $this->getUrlSuffix()) ?: $text;
+    }
+
+    public function setText(?string $text = null): void
+    {
+        $this->_text = $text;
     }
 
     public function getTarget(): ?string
@@ -449,11 +461,17 @@ abstract class Link extends Element implements LinkInterface
             return null;
         }
 
-        // Remove overridden `linkText` or `text` and use as the text
-        $customText = ArrayHelper::remove($attributes, 'linkText') ?? ArrayHelper::remove($attributes, 'text');
+        // Allow custom overriding of `text` for the label
+        if ($customText = ArrayHelper::remove($attributes, 'text')) {
+            $this->text = $customText;
+        }
 
-        if ($customText) {
+        // `linkText` shouldn't be overridden unless you know what you're doing, as this is the text defined
+        // in the field. It's here for backwards compatibility, but should be removed at the next breakpoint.
+        if ($customText = ArrayHelper::remove($attributes, 'linkText')) {
             $this->linkText = $customText;
+
+            Craft::$app->getDeprecator()->log(__METHOD__, 'Setting a link’s `linkText` has been deprecated. Use `text` instead.');
         }
 
         $attributes = $this->getLinkAttributes($attributes);