+++ src/fieldlayoutelements/CustomAttributesField.php
@@ -17,6 +17,16 @@ class CustomAttributesField extends BaseNativeField
 
     // Public Methods
     // =========================================================================
+
+    public function __construct($config = [])
+    {
+        unset(
+            $config['mandatory'],
+            $config['autofocus']
+        );
+
+        parent::__construct($config);
+    }
     
     public function showAttribute(): bool
     {