+++ src/fieldlayoutelements/LinkField.php
@@ -17,6 +17,7 @@ class LinkField extends TextField
     public ?HyperField $field = null;
     public ?LinkInterface $link = null;
     public bool $mandatory = true;
+    public bool $requirable = true;
 
 
     // Public Methods
@@ -26,7 +27,9 @@ class LinkField extends TextField
     {
         unset(
             $config['mandatory'],
+            $config['autofocus'],
             $config['linkType'],
+            $config['requirable'],
         );
 
         parent::__construct($config);