+++ src/fields/HyperField.php
@@ -600,7 +600,15 @@ class HyperField extends Field
             // because the form HTML is for JavaScript; not returned by inputHtml().
             return $view->namespaceInputs($form->render());
         } catch (Throwable $e) {
-            return Html::tag('div', Craft::t('hyper', 'Unable to render field - {e}.', ['e' => $e->getMessage()]), ['class' => 'error']);
+            $error = Craft::t('hyper', 'Unable to render field - {message} {file}:{line}', [
+                'message' => $e->getMessage(),
+                'file' => $e->getFile(),
+                'line' => $e->getLine(),
+            ]);
+
+            Hyper::error($error);
+
+            return Html::tag('div', $error, ['class' => 'error']);
         }
     }