+++ src/migrations/MigrateTypedLinkField.php
@@ -144,8 +144,16 @@ class MigrateTypedLinkField extends PluginFieldMigration
                     if ($matrixField) {
                         $this->migrateBlockField($matrixField, $newField);
 
+                        // For complex fields like Matrix > ST > Matrix, check if this is the top-level
                         if (!$fieldService->saveField($matrixField)) {
-                            throw new Exception(Json::encode($matrixField->getErrors()));
+                            $errors = $matrixField->getErrors();
+
+                            // Check for blocktype errors too
+                            foreach ($matrixField->getBlockTypes() as $blockType) {
+                                $errors[$blockType->handle] = $blockType->getErrors();
+                            }
+
+                            throw new Exception(Json::encode(array_filter($errors)));
                         }
                     } else {
                         $this->stdout("    > Unable to find owner Matrix field for ID “{$matrixFieldId}”." . PHP_EOL, Console::FG_RED);
@@ -171,8 +179,16 @@ class MigrateTypedLinkField extends PluginFieldMigration
                     if ($superTableField) {
                         $this->migrateBlockField($superTableField, $newField);
 
+                        // For complex fields like Matrix > ST > Matrix, check if this is the top-level
                         if (!$fieldService->saveField($superTableField)) {
-                            throw new Exception(Json::encode($superTableField->getErrors()));
+                            $errors = $superTableField->getErrors();
+
+                            // Check for blocktype errors too
+                            foreach ($superTableField->getBlockTypes() as $blockType) {
+                                $errors[] = $blockType->getErrors();
+                            }
+
+                            throw new Exception(Json::encode(array_filter($errors)));
                         }
                     } else {
                         $this->stdout("    > Unable to find owner Super Table field for ID “{$superTableFieldId}”." . PHP_EOL, Console::FG_RED);